Integrate the 8x resolution FLL into the regular timing loop

+++ timing.h
@@ -7,7 +7,7 @@ extern int32 make_ns(unsigned char ints, unsigned short counter);
 extern int32 time_get_ns();
 extern void time_get_ns_capt();
 extern void tickadj_set_ppm(signed short ppm);
-extern void tickadj_set_clocks(signed short clocks);
+extern void tickadj_set_clocks(int32 clocks);
 extern void pll_run();
 
 extern void time_set_date(unsigned int gps_week, uint32 gps_tow_sec, int offset);
